Move tests to replacer too
